chore: feature flag for profile page
