+++ src/app/[lng]/(infos)/profile/[walletAddress]/page.tsx
@@ -1,10 +1,16 @@
-import ProfilePage from '@/app/ui/profile/ProfilePage';
 import { getSiteUrl } from '@/const/urls';
 import { walletAddressSchema } from '@/utils/validation-schemas';
 import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
+import { Suspense } from 'react';
+import { getFeatureFlag } from 'src/app/lib/getFeatureFlag';
 import { getProfileBannerCampaigns } from 'src/app/lib/getProfileBannerCampaigns';
 import { getQuestsWithNoCampaignAttached } from 'src/app/lib/getQuestsWithNoCampaignAttached';
+import OldProfilePage from 'src/app/ui/profile/OldProfilePage';
+import { ProfilePage } from 'src/components/ProfilePage/ProfilePage';
+import { ProfilePageSkeleton } from 'src/components/ProfilePage/ProfilePageSkeleton';
+import { GlobalFeatureFlags } from 'src/const/abtests';
+import { fetchQuestOpportunitiesByRewardsIds } from 'src/utils/merkl/fetchQuestOpportunities';
 
 type Params = Promise<{ walletAddress: string }>;
 
@@ -61,14 +67,40 @@ export default async function Page({ params }: { params: Params }) {
   }
 
   const sanitizedAddress = result.data;
-  const { data: campaigns } = await getProfileBannerCampaigns();
-  const { data: questsData } = await getQuestsWithNoCampaignAttached();
+  const [{ data: campaigns }, { data: questsData }, isPageEnabled] =
+    await Promise.all([
+      getProfileBannerCampaigns(),
+      getQuestsWithNoCampaignAttached(),
+      getFeatureFlag(
+        GlobalFeatureFlags.ProfilePage,
+        // Placeholder distinctId required by the API call.
+        // This global feature flag is not tied to any specific user.
+        'distinct-id',
+      ),
+    ]);
+
+  const questsExtended = await fetchQuestOpportunitiesByRewardsIds(
+    questsData.data,
+  );
+
+  if (isPageEnabled) {
+    return (
+      <Suspense fallback={<ProfilePageSkeleton />}>
+        <ProfilePage
+          isPublic={true}
+          campaigns={campaigns}
+          quests={questsExtended}
+        />
+      </Suspense>
+    );
+  }
+
   return (
-    <ProfilePage
-      quests={questsData.data}
+    <OldProfilePage
+      quests={questsExtended}
       campaigns={campaigns}
       walletAddress={sanitizedAddress}
-      isPublic
+      isPublic={false}
     />
   );
 }
