+++ src/app/[lng]/(infos)/profile/page.tsx
@@ -1,8 +1,13 @@
 import { getProfileBannerCampaigns } from '@/app/lib/getProfileBannerCampaigns';
-import ProfilePage from '@/app/ui/profile/ProfilePage';
 import { getSiteUrl } from '@/const/urls';
 import type { Metadata } from 'next';
+import { Suspense } from 'react';
+import { getFeatureFlag } from 'src/app/lib/getFeatureFlag';
 import { getQuestsWithNoCampaignAttached } from 'src/app/lib/getQuestsWithNoCampaignAttached';
+import OldProfilePage from 'src/app/ui/profile/OldProfilePage';
+import { ProfilePage } from 'src/components/ProfilePage/ProfilePage';
+import { ProfilePageSkeleton } from 'src/components/ProfilePage/ProfilePageSkeleton';
+import { GlobalFeatureFlags } from 'src/const/abtests';
 import { fetchQuestOpportunitiesByRewardsIds } from 'src/utils/merkl/fetchQuestOpportunities';
 
 export const metadata: Metadata = {
@@ -15,15 +20,40 @@ export const metadata: Metadata = {
 };
 
 export default async function Page() {
-  const [{ data: campaigns }, { data: questsData }] = await Promise.all([
-    getProfileBannerCampaigns(),
-    getQuestsWithNoCampaignAttached(),
-  ]);
+  const [{ data: campaigns }, { data: questsData }, isPageEnabled] =
+    await Promise.all([
+      getProfileBannerCampaigns(),
+      getQuestsWithNoCampaignAttached(),
+      getFeatureFlag(
+        GlobalFeatureFlags.ProfilePage,
+        // Placeholder distinctId required by the API call.
+        // This global feature flag is not tied to any specific user.
+        'distinct-id',
+      ),
+    ]);
 
   // Fetch max APY for all quests and add to quest data
   const questsExtended = await fetchQuestOpportunitiesByRewardsIds(
     questsData.data,
   );
 
-  return <ProfilePage quests={questsExtended} campaigns={campaigns} />;
+  if (isPageEnabled) {
+    return (
+      <Suspense fallback={<ProfilePageSkeleton />}>
+        <ProfilePage
+          isPublic={true}
+          campaigns={campaigns}
+          quests={questsExtended}
+        />
+      </Suspense>
+    );
+  }
+
+  return (
+    <OldProfilePage
+      isPublic={false}
+      quests={questsExtended}
+      campaigns={campaigns}
+    />
+  );
 }
