+++ src/components/ProfilePage/ProfilePage.tsx
@@ -1,95 +1,35 @@
 'use client';
-import { useLoyaltyPass } from '@/hooks/useLoyaltyPass';
-import type { CampaignData, StrapiMediaData } from '@/types/strapi';
-import { useContext } from 'react';
-import { useTraits } from 'src/hooks/useTraits';
-import { AddressCard } from './AddressCard/AddressCard';
-import { LeaderboardCard } from './LeaderboardCard/LeaderboardCard';
-import { TierBox } from './LevelBox/TierBox';
-import {
-  PageContainer,
-  ProfileHeaderBox,
-  ProfileInfoBox,
-} from './ProfilePage.style';
-import { QuestsCompletedCarousel } from './QuestsCompletedCarousel/QuestsCompletedCarousel';
-import { QuestsOverview } from './QuestsOverview/QuestsOverview';
-// import { useABTest } from 'src/hooks/useABTest';
-
-import { MerklRewards } from '@/components/ProfilePage/MerklRewards';
-import { ProfileContext } from '@/providers/ProfileProvider';
-import { useMerklRewards } from 'src/hooks/useMerklRewards';
+import { useAccount } from '@lifi/wallet-management';
+import { ProfileProvider } from 'src/providers/ProfileProvider';
 import { QuestDataExtended } from 'src/types/merkl';
+import { CampaignData } from 'src/types/strapi';
+import { PageContainer } from '../Containers/PageContainer';
 
-import { CampaignBanners } from './CampaignBanners/CampaignBanners';
-import { IntroSection } from './sections/IntroSection';
-import { RewardsSection } from './sections/RewardsSection';
 interface ProfilePageProps {
+  walletAddress?: string;
+  isPublic?: boolean;
   campaigns?: CampaignData[];
   quests?: QuestDataExtended[];
-  questApys?: Record<string, number>;
 }
 
-// Type guard to filter campaigns that can be displayed in banners
-export interface CampaignWithBanner extends CampaignData {
-  ProfileBannerImage: StrapiMediaData;
-  ProfileBannerTitle: string;
-  ProfileBannerDescription: string;
-  ProfileBannerBadge: string;
-  ProfileBannerCTA?: string;
-  Slug: string;
-}
-
-const isBannerCampaign = (
-  campaign: CampaignData,
-): campaign is CampaignWithBanner =>
-  Boolean(
-    campaign.ProfileBannerImage?.url &&
-      campaign.ProfileBannerTitle &&
-      campaign.ProfileBannerDescription &&
-      campaign.ProfileBannerBadge &&
-      campaign.Slug,
-  );
-
-export const ProfilePage = ({ campaigns, quests }: ProfilePageProps) => {
-  const { walletAddress, isPublic } = useContext(ProfileContext);
-  const { isLoading, points, pdas } = useLoyaltyPass(walletAddress);
-  const { traits } = useTraits();
-
-  // const { isEnabled: isABTestEnabled } = useABTest({
-  //   feature: 'test_ab_1',
-  //   user: account?.address || '',
-  // });
-
-  const { pastCampaigns } = useMerklRewards({
-    userAddress: walletAddress,
-  });
-
-  const validBannerCampaigns = campaigns?.filter(isBannerCampaign) || [];
+export const ProfilePage = ({
+  walletAddress,
+  isPublic,
+  campaigns,
+  quests,
+}: ProfilePageProps) => {
+  const { account } = useAccount();
 
   return (
-    <PageContainer className="profile-page">
-      {!isPublic && <MerklRewards />}
-      {/* <IntroSection /> */}
-      {/* {!isPublic && <RewardsSection />} */}
-      <ProfileHeaderBox>
-        <AddressCard address={walletAddress} />
-        <ProfileInfoBox sx={{ display: 'flex', flex: 2, gap: 2 }}>
-          <TierBox points={points} loading={isLoading} />
-          <LeaderboardCard address={walletAddress} />
-        </ProfileInfoBox>
-      </ProfileHeaderBox>
-      {Array.isArray(validBannerCampaigns) &&
-        validBannerCampaigns.length > 0 && (
-          <CampaignBanners campaigns={validBannerCampaigns} />
-        )}
-      {Array.isArray(quests) && quests?.length > 0 && (
-        <QuestsOverview
-          quests={quests}
-          pastCampaigns={pastCampaigns}
-          traits={traits}
-        />
-      )}
-      <QuestsCompletedCarousel pdas={pdas} loading={isLoading} />
-    </PageContainer>
+    <ProfileProvider
+      walletAddress={walletAddress || account?.address || ''}
+      isPublic={isPublic}
+      // @Note these flags are not correctly set in @lifi/wallet-management
+      isLoading={account?.isConnecting || account?.isReconnecting}
+    >
+      <PageContainer>
+        <p>this is the new profile page</p>
+      </PageContainer>
+    </ProfileProvider>
   );
 };
