+++ src/components/ProfilePage/ProfilePageSkeleton.tsx
@@ -1,0 +1,3 @@
+export const ProfilePageSkeleton = () => {
+  return <div>This is the new profile page skeleton</div>;
+};
