+++ src/components/ProfilePage/index.ts
@@ -1,2 +1,3 @@
 export * from './AddressCard/IconButton.style';
+export * from './OldProfilePage';
 export * from './ProfilePage';
