+++ src/const/abtests.ts
@@ -20,4 +20,5 @@ export const AbTestConfig = {
 
 export enum GlobalFeatureFlags {
   MissionsPage = 'missions_page',
+  ProfilePage = 'profile_page',
 }
